+++ packages-automation/auto-invoiceAlert/src/createInvoiceAlertFromContracts.ts
@@ -1,61 +1,66 @@
-import { filterContractsToAlertTarget } from './helpers/filterContractsToAlertTarget';
-import { convertContractsToJson } from './helpers/convertContractsToJson';
+import { registerReminders } from './helpers/registerReminders';
+import { GetMyOrdersResponse } from 'api-andpad';
 import { ContractRecordType, IInvoiceReminder } from '../config';
 import { IAndpadpayments, IEmployees, IProjects, IStores } from 'types';
-//import { getTargetAndpadOrders } from './helpers/getTargetAndpadOrders';
-import { GetMyOrdersResponse } from 'api-andpad';
+import { filterContractsToAlertTarget } from './helpers/filterContractsToAlertTarget';
+import { convertContractsToJson } from './helpers/convertContractsToJson';
 
 
 
 /**
- * 契約アプリからアラート対象レコードを取得する
+ * 通知対象の契約レコード情報をまとめて
+ * kintoneのリマインダーアプリへ登録します
  */
-export const createInvoiceAlertFromContracts = ({
-  tgtProjTypeContracts,
-  andpadPayments,
-  reminders,
-  projects,
-  employees,
-  stores,
+export const createInvoiceAlertFromContracts = async ({
   allOrders,
+  allAndpadPayments,
+  allMembers,
+  allProjects,
+  allInvoiceReminder,
+  allStores,
+  tgtProjTypeContracts,
   allContracts,
 }: {
   tgtProjTypeContracts: ContractRecordType[]
-  andpadPayments: IAndpadpayments[]
-  reminders: IInvoiceReminder[]
-  projects: IProjects[]
-  employees: IEmployees[]
-  stores: IStores[]
+  allAndpadPayments: IAndpadpayments[]
+  allInvoiceReminder: IInvoiceReminder[]
+  allProjects: IProjects[]
+  allMembers: IEmployees[]
+  allStores: IStores[]
   allOrders: GetMyOrdersResponse
   allContracts: ContractRecordType[]
 }) => {
 
+  // 契約書の内容からアラート対象を取得する
   const alertContracts = filterContractsToAlertTarget({
     contracts: tgtProjTypeContracts,
-    andpadPayments: andpadPayments,
-    reminders: reminders,
-    projects: projects,
+    andpadPayments: allAndpadPayments,
+    reminders: allInvoiceReminder,
+    projects: allProjects,
   });
 
-  const consoleContracts = alertContracts.map(({ projName }) => projName.value);
-  console.log('通知対象の契約', alertContracts.length, consoleContracts);
+  const consoleContracts1 = alertContracts.map(({ projName }) => projName.value);
+  console.log('通知対象の契約', alertContracts.length, consoleContracts1);
 
 
-  // ANDPAD案件一覧の取得処理
-  /* const tgtOrders = await getTargetAndpadOrders({
-    contracts: alertContracts,
-    projects: projects,
-  }); */
-
-
-  return convertContractsToJson({
+  const alertContractsJson = convertContractsToJson({
     contracts: alertContracts,
     allContracts: allContracts,
-    projects: projects,
-    employees: employees,
-    stores: stores,
+    projects: allProjects,
+    employees: allMembers,
+    stores: allStores,
     allOrders: allOrders,
   });
 
+
+  const consoleContracts = alertContractsJson.map(({ projName }) => projName);
+  console.log('通知対象の契約:絞り込み後', alertContractsJson.length, consoleContracts);
+
+  //throw new Error('アラート対象の抽出が完了しました');
+
+  // 契約書から取得したアラート用データをリマインダーアプリへ登録する
+  await registerReminders({
+    reminderJson: alertContractsJson,
+  });
 
 };
