+++ packages-automation/auto-invoiceAlert/src/helpers/updateReportedReminders.test.ts
@@ -3,6 +3,7 @@ import fs from 'fs';
 import path from 'path';
 import { InvoiceReminder } from '../../types/InvoiceReminder';
 import { updateReportedReminders } from './updateReportedReminders';
+import { getInvoiceRemindersByAlertDate } from '../api-kintone';
 
 
 describe('updateReportedReminders', () => {
@@ -12,8 +13,11 @@ describe('updateReportedReminders', () => {
     const invoiceAlertPath = path.join(__dirname, '../__TEST__/createInvoiceAlert.json');
     const reminderDat = JSON.parse(fs.readFileSync(invoiceAlertPath, 'utf8')) as InvoiceReminder[];
 
+    const alertReminder = await getInvoiceRemindersByAlertDate(new Date());
+
     await updateReportedReminders({
       reportedReminder: reminderDat,
+      existedReminder: alertReminder,
     });
 
     console.log('kintoneアプリでテスト結果を確認してください');
