+++ packages-automation/auto-invoiceAlert/src/invoiceReminder.ts
@@ -1,6 +1,11 @@
-import { createInvoiceAlert } from './createInvoiceAlert';
+import { getAllAndpadPayments, getAllContracts, getAllProjects, getAllStores, getEmployees } from 'api-kintone';
+import { getAllInvoiceReminder, getInvoiceRemindersByAlertDate } from './api-kintone';
+import { createInvoiceAlertFromContracts } from './createInvoiceAlertFromContracts';
 import { updateReportedReminders } from './helpers/updateReportedReminders';
 import { notifyInvoiceAlertToChatwork } from './notifyInvoiceAlertToChatwork';
+import { filterContractsByTargetProjType } from './helpers/filterContractsByTargetProjType';
+import { getAllAndpadOrders } from 'api-andpad';
+import { convertReminderToJson } from './helpers/convertReminderToJson';
 
 
 /**
@@ -9,9 +14,57 @@ import { notifyInvoiceAlertToChatwork } from './notifyInvoiceAlertToChatwork';
 export const invoiceReminder = async () => {
   console.log('start invoice reminder');
 
-  const reminderJson = await createInvoiceAlert();
+  // 処理前準備
+  // 関連するレコード情報を取得する
+  const [
+    allProjects,
+    allAndpadPayments,
+    allMembers,
+    allStores,
+    tgtProjTypeContracts,
+    allInvoiceReminder,
+    allContracts,
+    allOrders,
+  ] = await Promise.all([
+    getAllProjects(),
+    getAllAndpadPayments(),
+    getEmployees(),
+    getAllStores(),
+    filterContractsByTargetProjType(),
+    getAllInvoiceReminder(),
+    getAllContracts(),
+    getAllAndpadOrders({ beforeInvoiceIssue: true }),
+  ]);
+
+
+  await createInvoiceAlertFromContracts({
+    allAndpadPayments: allAndpadPayments,
+    allContracts: allContracts,
+    allInvoiceReminder: allInvoiceReminder,
+    allMembers: allMembers,
+    allOrders: allOrders,
+    allProjects: allProjects,
+    allStores: allStores,
+    tgtProjTypeContracts: tgtProjTypeContracts,
+  });
+
+
+  // 今日通知予定のリマインダーレコードを取得する(含：契約書から取得したアラート)
+  const alertReminder = await getInvoiceRemindersByAlertDate(new Date());
+
+
+  const reminderJson = convertReminderToJson({
+    reminder: alertReminder,
+    andpadPayments: allAndpadPayments,
+    allOrders: allOrders,
+    allProjects: allProjects,
+    employees: allMembers,
+    stores: allStores,
+  });
+
 
   //throw new Error('抽出処理のみ完了');
+
 
   // chatworkへの通知処理
   notifyInvoiceAlertToChatwork({
@@ -21,6 +74,7 @@ export const invoiceReminder = async () => {
   // リマインダーレコードの更新処理
   updateReportedReminders({
     reportedReminder: reminderJson,
+    existedReminder: alertReminder,
   });
 
   console.log('finish invoice reminder');
